Split the api package into multiple files
