+++ api/wiki.go
@@ -1,0 +1,12 @@
+package api
+
+import gowiki "github.com/trietmn/go-wiki"
+
+func SearchWiki(word string) (string, error) {
+	res, err := gowiki.Summary(word, 5, -1, false, true)
+	if err != nil {
+		return "", err
+	}
+
+	return res, nil
+}
